requirements.txt
requests
fastapi
mangum
pymongo
python-dotenv
python-jose
python-multipart
bcrypt